Add styles to the component. The component is responsive
