+++ src/js/index.js
@@ -3,7 +3,7 @@ import { defaultComp } from './components/default';
 import { navigationComponent } from './components/navigation';
 import { scrollIndicator } from './components/scroll-indicator';
 import { programsComp } from './components/programs';
-import { escenariosComp } from './components/escenarios';
+import { scenariosComp } from './components/scenarios';
 import { eventsCarousel } from './components/events-carousel';
 import { footerComp } from './components/footer';
 import { pageData } from '../data/data';
@@ -17,7 +17,7 @@ const components = {
   navigation: navigationComponent,
   'scroll-indicator': scrollIndicator,
   programs: programsComp,
-  scenarios: escenariosComp,
+  scenarios: scenariosComp,
   events: eventsCarousel,
   footer: footerComp,
 };
